+++ CHANGELOG.md
@@ -14,7 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - rooks and bishops valid_postions updated due to the change in 'Added' 
-- simplified all moves method
+- simplified all moves 
+- simplified possible_positions_vertical_and_horizontal
 
 ## [0.2.2] - 2024-10-17