+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - simplified possible_positions_vertical_and_horizontal
 - simplified possible_positions_diagonal
 - divided all_moves in more methods to pass codeclimate
+- updated knight moves
 
 ## [0.2.2] - 2024-10-17