+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - simplified possible_positions_diagonal
 - divided all_moves in more methods to pass codeclimate
 - updated knight moves
+- updated board str
 
 ## [0.2.2] - 2024-10-17