+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - valid_positions_rook_and_bishop unites the 2 valid positions in one method
 - Rules function in chess to use in the cli to verifier rules after every move
+- get_possible_positions method
 
 ### Changed
 
@@ -21,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - updated knight moves
 - updated board str
 - updated pawn change to other piece method
+- changed structur for valid-postions in king an knight, now uses get_possible_postiosn function
 
 ## [0.2.2] - 2024-10-17